+++ app/assets/javascripts/discourse/app/helpers/user-avatar.js
@@ -1,4 +1,5 @@
-import { avatarImg, formatUsername } from "discourse/lib/utilities";
+import { formatUsername } from "discourse/lib/utilities";
+import { avatarImg } from "discourse-common/lib/avatar-utils";
 import I18n from "I18n";
 import { get } from "@ember/object";
 import { htmlSafe } from "@ember/template";
