+++ lib/pretty_text/shims.js
@@ -24,8 +24,6 @@ define("discourse-common/lib/helpers", ["exports"], function (exports) {
   };
 });
 
-__utils = require("discourse/lib/utilities");
-
 __emojiUnicodeReplacer = null;
 
 __setUnicode = function (replacements) {
@@ -119,7 +117,7 @@ function __hashtagLookup(slug, cookingUserId, typesInPriorityOrder) {
 }
 
 function __lookupAvatar(p) {
-  return __utils.avatarImg(
+  return require("discourse-common/lib/avatar-utils").avatarImg(
     { size: "tiny", avatarTemplate: __helpers.avatar_template(p) },
     __getURL
   );
